refactor: OrderTables 패키지 분리 및 common 패키지 분리
