+++ src/main/java/kitchenpos/ordertable/domain/OrderTables.java
@@ -1,6 +1,4 @@
-package kitchenpos.domain;
-
-import kitchenpos.ordertable.domain.OrderTable;
+package kitchenpos.ordertable.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
