+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -3,9 +3,9 @@ package kitchenpos.tablegroup.application;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.OrderTables;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTables;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.request.TableGroupUnitDto;
 import kitchenpos.tablegroup.domain.TableGroup;
