+++ src/test/java/kitchenpos/common/OrderTablesTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.common;
 
 import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTables;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
